Add a default timeout on socket objects
Currently sockets aren't created with timeouts and the service
occasionally hangs. This doesn't fix the underlying bug, but it should
ensure the service never gets permanently stuck waiting on socket
operations. The value of the timeout was chosen arbitrarily and is
hopefully absurdly high. Some packages take a very long time to sign so
we might need to bump this.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>