sign_pe: Add the ability to sign PE applications via PKCS#11
This adds a new key type, PKCS#11 (yes, not _really_ a type, but there's
already a gnupg key type) and adds a check for the key type in the
sign_pe command.

Additionally, there's a helper script to add a row into the keys
table for each PKCS#11-backed key you wish to sign with.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>