db: Add a database migration for pkcs11 key type
The v11 keys table introduces a constraint on the keytype to be one of
'gnupg' or 'ECC'. There is, however, an additional key type of 'RSA'
which is required when using pesign, and the HSM support introduces a
further key type of 'PKCS11' (yes, not really a type).

Introduce the new 'PKCS11' type to the keytype column constraint, as
well as add in the forgotten 'RSA' type.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>